Add timeseries display to topology view
Restrucute the spatial display such that the topology view and the spatial dipslay view can use the same component
